display: decrease DSR_COUNT in order to sleep more
DSR count had been increased from 2 to 15 three years ago in order to make
it less power-agressive and avoid shutting down the display for nothing.
On our platforms, we need to reduce it in order to increase the s0ix residency
in use cases with a few fps. It will allow to sleep more between 2 frames.
As free_count is incremented every 16ms, 2 is maybe a bit too agressive (32ms)
but 4 looks like a good trade-off and should be safe enough.

Change-Id: Ie2e142fb7e25897d144c352e6eb3f2f7d0dd8573
Tracked-On: https://jira01.devtools.intel.com/browse/AW-4777
Signed-off-by: Simon Dubray <simonx.dubray@intel.com>
Reviewed-on: https://android.intel.com/565718
Reviewed-by: Lachaud, EtienneX <etiennex.lachaud@intel.com>
Reviewed-by: Bel Aj Ali, HabibX <habibx.bel.aj.ali@intel.com>
Tested-by: Bel Aj Ali, HabibX <habibx.bel.aj.ali@intel.com>
Reviewed-by: Tasayco Loarte, VictorX <victorx.tasayco.loarte@intel.com>
Reviewed-on: https://android.intel.com/569082
Reviewed-by: jenkins_ndg <jenkins_ndg@intel.com>
Reviewed-by: Deverge, Jean-francoisX <jean-francoisx.deverge@intel.com>
Tested-by: Deverge, Jean-francoisX <jean-francoisx.deverge@intel.com>
Reviewed-by: Whitfield, MichaelX <michaelx.whitfield@intel.com>
Reviewed-by: Jacquet, CyrilX <cyrilx.jacquet@intel.com>
Reviewed-by: Ledentec, AlexandreX <alexandrex.ledentec@intel.com>